initial impl of tileset json

+++ nusamai-gltf/nusamai-gltf-json/src/models/camera.rs
@@ -9,16 +9,16 @@ use serde_json::Value;
 #[serde(deny_unknown_fields)]
 pub struct CameraOrthographic {
     /// The floating-point horizontal magnification of the view. This value MUST NOT be equal to zero. This value SHOULD NOT be negative.
-    pub xmag: f32,
+    pub xmag: f64,
 
     /// The floating-point vertical magnification of the view. This value MUST NOT be equal to zero. This value SHOULD NOT be negative.
-    pub ymag: f32,
+    pub ymag: f64,
 
     /// The floating-point distance to the far clipping plane. This value MUST NOT be equal to zero. zfar MUST be greater than znear.
-    pub zfar: f32,
+    pub zfar: f64,
 
     /// The floating-point distance to the near clipping plane.
-    pub znear: f32,
+    pub znear: f64,
 }
 
 /// A perspective camera containing properties to create a perspective projection matrix.
@@ -28,17 +28,17 @@ pub struct CameraOrthographic {
 pub struct CameraPerspective {
     /// The floating-point aspect ratio of the field of view. When undefined, the aspect ratio of the rendering viewport MUST be used.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub aspect_ratio: Option<f32>,
+    pub aspect_ratio: Option<f64>,
 
     /// The floating-point vertical field of view in radians. This value SHOULD be less than π.
-    pub yfov: f32,
+    pub yfov: f64,
 
     /// The floating-point distance to the far clipping plane. When defined, zfar MUST be greater than znear. If zfar is undefined, client implementations SHOULD use infinite projection matrix.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub zfar: Option<f32>,
+    pub zfar: Option<f64>,
 
     /// The floating-point distance to the near clipping plane.
-    pub znear: f32,
+    pub znear: f64,
 }
 
 #[derive(Serialize, Deserialize, Debug, Clone, Copy, Default)]
