+++ nusamai-gltf/nusamai-gltf-json/src/models/material.rs
@@ -31,7 +31,7 @@ pub struct MaterialNormalTextureInfo {
 
     /// The scalar parameter applied to each normal vector of the texture. This value scales the normal vector in X and Y directions using the formula: `scaledNormal =  normalize((<sampled normal texture value> * 2.0 - 1.0) * vec3(<normal scale>, <normal scale>, 1.0))`."""
     #[serde(default = "one", skip_serializing_if = "is_one")]
-    pub scale: f32,
+    pub scale: f64,
 
     /// JSON object with extension-specific objects.
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -74,7 +74,7 @@ pub struct MaterialOcclusionTextureInfo {
 
     /// A scalar parameter controlling the amount of occlusion applied. A value of `0.0` means no occlusion. A value of `1.0` means full occlusion. This value affects the final occlusion value as: `1.0 + strength * (<sampled occlusion texture value> - 1.0)`.
     #[serde(default = "one", skip_serializing_if = "is_one")]
-    pub strength: f32,
+    pub strength: f64,
 
     /// JSON object with extension-specific objects.
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -112,7 +112,7 @@ pub struct MaterialOcculusionTextureInfoExtensions {
 pub struct MaterialPbrMetallicRoughness {
     /// The factors for the base color of the material. This value defines linear multipliers for the sampled texels of the base color texture.
     #[serde(skip_serializing_if = "is_default_color")]
-    pub base_color_factor: [f32; 4],
+    pub base_color_factor: [f64; 4],
 
     /// The base color texture. The first three components (RGB) **MUST** be encoded with the sRGB transfer function. They specify the base color of the material. If the fourth component (A) is present, it represents the linear alpha coverage of the material. Otherwise, the alpha coverage is equal to `1.0`. The `material.alphaMode` property specifies how alpha is interpreted. The stored texels **MUST NOT** be premultiplied. When undefined, the texture **MUST** be sampled as having `1.0` in all components.
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -120,18 +120,18 @@ pub struct MaterialPbrMetallicRoughness {
 
     /// The factor for the metalness of the material. This value defines a linear multiplier for the sampled metalness values of the metallic-roughness texture.
     #[serde(skip_serializing_if = "is_one")]
-    pub metallic_factor: f32,
+    pub metallic_factor: f64,
 
     /// The factor for the metalness of the material. This value defines a linear multiplier for the sampled metalness values of the metallic-roughness texture.
     #[serde(skip_serializing_if = "is_one")]
-    pub roughness_factor: f32,
+    pub roughness_factor: f64,
 
     /// The metallic-roughness texture. The metalness values are sampled from the B channel. The roughness values are sampled from the G channel. These values **MUST** be encoded with a linear transfer function. If other channels are present (R or A), they **MUST** be ignored for metallic-roughness calculations. When undefined, the texture **MUST** be sampled as having `1.0` in G and B components.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub metallic_roughness_texture: Option<TextureInfo>,
 }
 
-fn is_default_color(v: &[f32; 4]) -> bool {
+fn is_default_color(v: &[f64; 4]) -> bool {
     *v == [1.0, 1.0, 1.0, 1.0]
 }
 
@@ -175,7 +175,7 @@ pub struct Material {
 
     /// The factors for the emissive color of the material. This value defines linear multipliers for the sampled texels of the emissive texture.
     #[serde(skip_serializing_if = "is_default_emissive_factor")]
-    pub emissive_factor: [f32; 3],
+    pub emissive_factor: [f64; 3],
 
     /// The material's alpha rendering mode enumeration specifying the interpretation of the alpha value of the base color.
     #[serde(skip_serializing_if = "is_default")]
@@ -183,7 +183,7 @@ pub struct Material {
 
     /// Specifies the cutoff threshold when in `MASK` alpha mode. If the alpha value is greater than or equal to this value then it is rendered as fully opaque, otherwise, it is rendered as fully transparent. A value greater than `1.0` will render the entire material as fully transparent. This value **MUST** be ignored for other alpha modes. When `alphaMode` is not defined, this value **MUST NOT** be defined.
     #[serde(skip_serializing_if = "is_half")]
-    pub alpha_cutoff: f32,
+    pub alpha_cutoff: f64,
 
     /// Specifies whether the material is double sided. When this value is false, back-face culling is enabled. When this value is true, back-face culling is disabled and double-sided lighting is enabled. The back-face **MUST** have its normals reversed before the lighting equation is evaluated.
     #[serde(skip_serializing_if = "is_default")]
@@ -223,19 +223,19 @@ pub struct MaterialExtensions {
     others: HashMap<String, Value>,
 }
 
-fn one() -> f32 {
+fn one() -> f64 {
     1.0
 }
 
-fn is_one(v: &f32) -> bool {
+fn is_one(v: &f64) -> bool {
     *v == 1.0
 }
 
-fn is_default_emissive_factor(v: &[f32; 3]) -> bool {
+fn is_default_emissive_factor(v: &[f64; 3]) -> bool {
     *v == [0.0, 0.0, 0.0]
 }
 
-fn is_half(v: &f32) -> bool {
+fn is_half(v: &f64) -> bool {
     *v == 0.5
 }
